Authorisation

+++ Web/Controllers/RoleController.cs
@@ -10,6 +10,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using ViewModels.Input;
 using Web.Services.Interfaces;
+using Repositories.Helpers;
 
 namespace Web.Controllers
 {
@@ -28,8 +29,13 @@ namespace Web.Controllers
 		[HttpGet]
 		public IActionResult Create()
 		{
-			RoleViewModel model = new RoleViewModel();
-			return View(model);
+			if(Logged.CEOAuth())
+            {
+				RoleViewModel model = new RoleViewModel();
+				return View(model);
+			}
+			return Unauthorized();
+			
 		}
 
 		[HttpPost]
@@ -53,13 +59,18 @@ namespace Web.Controllers
 		[Route("post/edit/{id}")]
 		public IActionResult Edit([FromRoute] string id)
 		{
-			if (string.IsNullOrEmpty(id))
+			if (Logged.CEOAuth())
 			{
-				return NotFound();
+				if (string.IsNullOrEmpty(id))
+				{
+					return NotFound();
+				}
+				EditRoleViewModel model = new EditRoleViewModel();
+				model.Id = id;
+				return View(model);
 			}
-			EditRoleViewModel model = new EditRoleViewModel();
-			model.Id = id;
-			return View(model);
+			else return Unauthorized();
+			
 		}
 
 		[HttpPost]
@@ -84,17 +95,22 @@ namespace Web.Controllers
 		[Route("role/delete/{id}")]
 		public IActionResult Delete([FromRoute] string id)
 		{
-			if (id is null)
+			if (Logged.CEOAuth())
 			{
-				return NotFound();
+				if (id is null)
+				{
+					return NotFound();
+				}
+				Role deleteRole = _roleService.GetRole(id);
+				if (deleteRole is null)
+				{
+					return NotFound();
+				}
+				_roleService.DeleteRole(deleteRole);
+				return RedirectToAction("Index", "Role");
 			}
-			Role deleteRole = _roleService.GetRole(id);
-			if (deleteRole is null)
-			{
-				return NotFound();
-			}
-			_roleService.DeleteRole(deleteRole);
-			return RedirectToAction("Index", "Role");
+			else return Unauthorized();
+			
 
 		}
 		#endregion
